Fix: onUpdated updates tab data only on navigation events.
The idea behind using onUpdated in the extension is to get informed
about navigation changes and resetting the reload timeout for that tab.

onUpdated gets much more called in recent versions than before thus the
tab data is updated way more often and falsy. In the end it could occur
often, that a tab does not get reloaded anymore.

So the extension looks now explicitly for navigation events.

Ticket #3